+++ CHANGELOG.md
@@ -8,6 +8,7 @@ Release:     1.7.0 (May 2020)
 KEY CHANGES:
 - [ENH] DXGI: Expose types from windows.ui.xaml.media.dxinterop.h
 - [FIX] WIC: Crash with IWICBitmapLock.GetDataPointer, correctly handle different types
+- [FIX] Direct3D11: Fix OMSetRenderTargetsAndUnorderedAccessViews overloads and add KeepRenderTargetsAndDepthStencil and KeepUnorderedAccessViews support.
 
 -----------------------------------------------
 Release:     1.6.0 (April 2020)